+++ R/expect.R
@@ -1,7 +1,7 @@
 
-#' \code{testthat} expectation for a Shiny update
+#' `testthat` expectation for a Shiny update
 #'
-#' @param app A \code{\link{ShinyDriver}} object.
+#' @param app A [ShinyDriver()] object.
 #' @param output Character vector, the name(s) of the output widgets
 #'   that are required to update for the test to succeed.
 #' @param ... Named arguments specifying updates for Shiny input
@@ -21,7 +21,7 @@
 #' expectUpdate(app, ycol = "Petal.Width", output = "plot1")
 #' expectUpdate(app, clusters = 4, output = "plot1")
 #' }
-
+#' @keywords internal
 expectUpdate <- function(app, output, ..., timeout = 3000,
                          iotype = c("auto", "input", "output")) {
   app$expectUpdate(