+++ R/initialize.R
@@ -4,27 +4,33 @@
 
 sd_initialize <- function(self, private, path, loadTimeout, checkNames,
                           debug, phantomTimeout, seed, cleanLogs,
-                          shinyOptions) {
+                          shinyOptions, renderArgs, options) {
 
   private$cleanLogs <- cleanLogs
+  if (is.null(loadTimeout)) {
+    loadTimeout <- if (on_ci()) 10000 else 5000
+  }
 
   self$logEvent("Start ShinyDriver initialization")
 
   if (is.null(find_phantom())) {
-    stop("PhantomJS not found.")
+    abort("PhantomJS not found.")
   }
 
   "!DEBUG get phantom port (starts phantom if not running)"
   self$logEvent("Getting PhantomJS port")
   private$phantomPort <- get_phantomPort(timeout = phantomTimeout)
 
+  if (is.shiny.appobj(path)) {
+    path <- app_save(path)
+  }
+
   if (grepl("^http(s?)://", path)) {
     private$setShinyUrl(path)
-
   } else {
     "!DEBUG starting shiny app from path"
     self$logEvent("Starting Shiny app")
-    private$startShiny(path, seed, loadTimeout, shinyOptions)
+    private$startShiny(path, seed, loadTimeout, shinyOptions, renderArgs, options)
   }
 
   "!DEBUG create new phantomjs session"
@@ -52,10 +58,10 @@ sd_initialize <- function(self, private, path, loadTimeout, checkNames,
     timeout = loadTimeout
   )
   if (!load_ok) {
-    stop(
+    abort(paste0(
       "Shiny app did not load in ", loadTimeout, "ms.\n",
       format(self$getDebugLog())
-    )
+    ))
   }
 
   "!DEBUG shiny started"
@@ -86,7 +92,7 @@ sd_initialize <- function(self, private, path, loadTimeout, checkNames,
 #' @importFrom rematch re_match
 #' @importFrom withr with_envvar
 
-sd_startShiny <- function(self, private, path, seed, loadTimeout, shinyOptions) {
+sd_startShiny <- function(self, private, path, seed, loadTimeout, shinyOptions, renderArgs, options) {
 
   assert_that(is_string(path))
 
@@ -104,7 +110,7 @@ sd_startShiny <- function(self, private, path, seed, loadTimeout, shinyOptions) 
   p <- with_envvar(
     c("R_TESTS" = NA),
     callr::r_bg(
-      function(path, shinyOptions, rmd, seed, rng_kind) {
+      function(path, shinyOptions, rmd, seed, rng_kind, renderArgs, options) {
 
         if (!is.null(seed)) {
           # Prior to R 3.6, RNGkind has 2 args, otherwise it has 3
@@ -113,17 +119,27 @@ sd_startShiny <- function(self, private, path, seed, loadTimeout, shinyOptions) 
           shiny:::withPrivateSeed(set.seed(seed + 11))
         }
 
-        options(shiny.testmode = TRUE)
+        options <- as.list(options)
+        options$shiny.testmode <- TRUE
+        do.call(base::options, options)
 
         if (rmd) {
           # Shiny document
-          rmarkdown::run(path, shiny_args = shinyOptions)
+          rmarkdown::run(path, shiny_args = shinyOptions, render_args = renderArgs)
         } else {
           # Normal shiny app
           do.call(shiny::runApp, c(path, shinyOptions))
         }
       },
-      args = list(path, shinyOptions, is_rmd(path), seed, rng_kind),
+      args = list(
+        path = path,
+        shinyOptions = shinyOptions,
+        rmd = is_rmd(path),
+        seed = seed,
+        rng_kind = rng_kind,
+        renderArgs = renderArgs,
+        options = options
+      ),
       stdout = sprintf(tempfile_format, "shiny-stdout"),
       stderr = sprintf(tempfile_format, "shiny-stderr"),
       supervise = TRUE
@@ -131,10 +147,10 @@ sd_startShiny <- function(self, private, path, seed, loadTimeout, shinyOptions) 
   )
   "!DEBUG waiting for shiny to start"
   if (! p$is_alive()) {
-    stop(
+    abort(paste0(
       "Failed to start shiny. Error: ",
       strwrap(readLines(p$get_error_file()))
-    )
+    ))
   }
 
   "!DEBUG finding shiny port"
@@ -144,14 +160,18 @@ sd_startShiny <- function(self, private, path, seed, loadTimeout, shinyOptions) 
     err_lines <- readLines(p$get_error_file())
 
     if (!p$is_alive()) {
-      stop("Error starting application:\n", paste(err_lines, collapse = "\n"))
+      abort(paste0(
+        "Error starting application:\n", paste(err_lines, collapse = "\n")
+      ))
     }
     if (any(grepl("Listening on http", err_lines))) break
 
     Sys.sleep(0.2)
   }
   if (i == max_i) {
-    stop("Cannot find shiny port number. Error:\n", paste(err_lines, collapse = "\n"))
+    abort(paste0(
+      "Cannot find shiny port number. Error:\n", paste(err_lines, collapse = "\n")
+    ))
   }
 
   line <- err_lines[grepl("Listening on http", err_lines)]
@@ -200,7 +220,7 @@ phantom_paths <- function() {
   if (is_windows()) {
     path <- Sys.getenv('APPDATA', '')
     path <- if (dir_exists(path)) file.path(path, 'PhantomJS')
-  } else if (is_osx()) {
+  } else if (is_mac()) {
     path <- '~/Library/Application Support'
     path <- if (dir_exists(path)) file.path(path, 'PhantomJS')
   } else {
@@ -208,36 +228,6 @@ phantom_paths <- function() {
   }
   path <- c(path, system.file('PhantomJS', package = 'webdriver'))
   path
-}
-
-# Find PhantomJS from PATH, APPDATA, system.file('webdriver'), ~/bin, etc
-find_phantom <- function(quiet = FALSE) {
-  path <- Sys.which( "phantomjs" )
-  if (path != "") return(path)
-
-  for (d in phantom_paths()) {
-    exec <- if (is_windows()) "phantomjs.exe" else "phantomjs"
-    path <- file.path(d, exec)
-    if (utils::file_test("-x", path)) break else path <- ""
-  }
-
-  if (path == "") {
-    if (!quiet) {
-      # It would make the most sense to throw an error here. However, that would
-      # cause problems with CRAN. The CRAN checking systems may not have phantomjs
-      # and may not be capable of installing phantomjs (like on Solaris), and any
-      # packages which use webdriver in their R CMD check (in examples or vignettes)
-      # will get an ERROR. We'll issue a message and return NULL; other
-      message(
-        "shinytest requires a headless web browser (PhantomJS) to record and run tests.\n",
-        "To install it, run shinytest::installDependencies()\n",
-        "If it is installed, please make sure the phantomjs executable ",
-        "can be found via the PATH variable."
-      )
-    }
-    return(NULL)
-  }
-  path.expand(path)
 }