+++ R/package.R
@@ -1,9 +1,0 @@
-
-#' Test Shiny Apps
-#'
-#' Uses a headless browser, driven through 'WebDriver'.
-#' See \code{\link{ShinyDriver}} to get started.
-#'
-#' @docType package
-#' @name shinytest
-NULL