+++ R/recorder.R
@@ -1,6 +1,6 @@
 #' Launch test event recorder for a Shiny app
 #'
-#' @param app A \code{\link{ShinyDriver}} object, or path to a Shiny
+#' @param app A [ShinyDriver()] object, or path to a Shiny
 #'   application.
 #' @param save_dir A directory to save stuff.
 #' @param load_mode A boolean that determines whether or not the resulting test
@@ -9,13 +9,13 @@
 #'   be used in the test script.
 #' @param loadTimeout Maximum time to wait for the Shiny application to load, in
 #'   milliseconds. If a value is provided, it will be saved in the test script.
-#' @param debug start the underlying \code{\link{ShinyDriver}} in \code{debug}
+#' @param debug start the underlying [ShinyDriver()] in `debug`
 #'   mode and print those debug logs to the R console once recording is
-#'   finished. The default, \code{'shiny_console'}, captures and prints R
+#'   finished. The default, `'shiny_console'`, captures and prints R
 #'   console output from the recorded R shiny process. Any value that the
-#'   \code{debug} argument in \code{\link{ShinyDriver}} accepts may be used
-#'   (e.g., \code{'none'} may be used to completely suppress the driver logs).
-#' @param shinyOptions A list of options to pass to \code{runApp()}. If a value
+#'   `debug` argument in [ShinyDriver()] accepts may be used
+#'   (e.g., `'none'` may be used to completely suppress the driver logs).
+#' @param shinyOptions A list of options to pass to `runApp()`. If a value
 #'   is provided, it will be saved in the test script.
 #' @export
 recordTest <- function(app = ".", save_dir = NULL, load_mode = FALSE, seed = NULL,
@@ -27,25 +27,16 @@ recordTest <- function(app = ".", save_dir = NULL, load_mode = FALSE, seed = NUL
     url <- app$getUrl()
   } else if (is.character(app)) {
     if (grepl("^http(s?)://", app)) {
-      stop("Recording tests for remote apps is not yet supported.")
+      abort("Recording tests for remote apps is not yet supported.")
     } else {
-      app <- app_path(app)
+      path <- app_path(app, "app")$app
 
-      if (is_rmd(app)) {
-        # If it's an Rmd file, make sure there aren't multiple Rmds in that
-        # directory.
-        if (length(dir(dirname(app), pattern = "\\.Rmd$", ignore.case = TRUE)) > 1) {
-          stop("For testing, only one .Rmd file is allowed per directory.")
-        }
-
-        # Rmds need a random seed. Automatically create one if needed.
-        if (is.null(seed)) {
-          seed <- floor(stats::runif(1, min = 0, max = 1e5))
-        }
+      # Rmds need a random seed
+      if (is_rmd(path) && is.null(seed)) {
+        seed <- floor(stats::runif(1, min = 0, max = 1e5))
       }
 
-      # It's a path to an app; start the app
-      app <- ShinyDriver$new(app, seed = seed, loadTimeout = loadTimeout, shinyOptions = shinyOptions)
+      app <- ShinyDriver$new(path, seed = seed, loadTimeout = loadTimeout, shinyOptions = shinyOptions)
       on.exit({
         rm(app)
         gc()
@@ -53,9 +44,9 @@ recordTest <- function(app = ".", save_dir = NULL, load_mode = FALSE, seed = NUL
       url <- app$getUrl()
     }
   } else if (inherits(app, "shiny.appobj")) {
-    stop("Recording tests for shiny.appobj objects is not supported.")
+    abort("Recording tests for shiny.appobj objects is not supported.")
   } else {
-    stop("Unknown object type to record tests for.")
+    abort("Unknown object type to record tests for.")
   }
 
   # Create directory if needed
@@ -116,13 +107,10 @@ recordTest <- function(app = ".", save_dir = NULL, load_mode = FALSE, seed = NUL
 
   } else {
     if (length(res$dont_run_reasons) > 0) {
-      message(
-        "Not running test script because:\n  ",
-        paste(res$dont_run_reasons, collapse = "\n  "), "\n"
-      )
+      inform(c("Not running test script because", res$dont_run_reasons))
     }
 
-    message(sprintf(
+    inform(sprintf(
       'After making changes to the test script, run it with:\n  testApp("%s", "%s")',
       rel_path(res$appDir), res$file
     ))
@@ -131,3 +119,34 @@ recordTest <- function(app = ".", save_dir = NULL, load_mode = FALSE, seed = NUL
   invisible(res$file)
 }
 
+
+#' Register an input processor for the test recorder
+#'
+#' @description
+#' `registerInputProcessor()` registers an input processor which will be used by
+#' the test recorder. The input processor function should take one parameter,
+#' `value`, and return a string of R code which returns the desired value.
+#'
+#' `getInputProcessors()` returns a named list of all registered input processor
+#' functions.
+#'
+#' @param inputType The name of an input type, for example,
+#'   `"mypkg.numberinput"`.
+#' @param processor An input processor function.
+#' @export
+#' @keywords internal
+registerInputProcessor <- function(inputType, processor) {
+  if (!is.function(processor) || !identical(names(formals(processor)), "value")) {
+    abort("`processor` must be a function that takes one parameter, `value`")
+  }
+  recorder_input_processors[[inputType]] <- processor
+}
+
+#' @rdname registerInputProcessor
+#' @export
+getInputProcessors <- function() {
+  as.list(recorder_input_processors)
+}
+
+# This environment holds input processors registered by other packages on load.
+recorder_input_processors <- new.env(parent = emptyenv())