+++ R/widget.R
@@ -1,151 +1,139 @@
-
-#' Class for a Shiny widget
+#' A Shiny Widget
 #'
-#' @section Usage:
-#' \preformatted{w <- app$findWidget(name,
-#'     iotype = c("auto", "input", "output"))
+#' @description
+#' A `Widget` object represents a Shiny input or output control, and provides
+#' methods for finer grained interaction.
 #'
-#' w$getName()
-#' w$getElement()
-#' w$getType()
-#' w$getIoType()
-#' w$isInput()
-#' w$isOutput()
-#'
-#' w$getValue()
-#' w$setValue(value)
-#'
-#' w$sendKeys(keys)
-#'
-#' w$listTabs()
-#' }
-#'
-#' @section Arguments:
-#' \describe{
-#'   \item{app}{A \code{\link{ShinyDriver}} object.}
-#'   \item{w}{A \code{Widget} object.}
-#'   \item{name}{Name of a Shiny widget.}
-#'   \item{iotype}{Character scalar, whether the widget is \sQuote{input}
-#'     or \sQuote{output}. The default \sQuote{auto} value works well,
-#'     provided that widgets have unique names. (Shiny allows an input
-#'     and an output widget with the same name.)}
-#'   \item{value}{Value to set for the widget. Its interpretation depends
-#'     on the type of the widget, see details below.}
-#'   \item{keys}{Keys to send to the widget. See the \code{sendKeys}
-#'     method of the \code{\link[webdriver]{Element}} class in the
-#'     \code{webdriver} package.}
-#' }
-#'
-#' @section Details:
-#'
-#' A \code{Widget} object represents a Shiny input or output widget.
-#' \code{app$findWidget} creates a widget object from a
-#' \code{\link{ShinyDriver}} object.
-#'
-#' \code{w$getName()} returns the name of the widget.
-#'
-#' \code{w$getElement()} returns an HTML element. This is an
-#' \code{\link[webdriver]{Element}} object from the \code{webdriver}
-#' package.
-#'
-#' \code{w$getType()} returns the type of the widget, possible values
-#' are \code{textInput}, \code{selectInput}, etc.
-#'
-#' \code{w$getIoType()} returns \sQuote{input} or \sQuote{output},
-#' whether the widget is an input or output widget.
-#'
-#' \code{w$isInput()} returns \code{TRUE} for input widgets, \code{FALSE}
-#' otherwise.
-#'
-#' \code{w$isOutput()} returns \code{TRUE} for output widgets, \code{FALSE}
-#' otherwise.
-#'
-#' \code{w$getValue()} returns the value of the widget. The exact type
-#' returned depends on the type of the widget. TODO: list widgets and their
-#' return types.
-#'
-#' \code{w$setValue()} sets the value of the widget, through the web
-#' browser. Different widget types expect different different \code{value}
-#' arguments. TODO: list widgets and types.
-#'
-#' \code{w$sendKeys} sends the specified keys to the HTML element of the
-#' widget.
-#'
-#' \code{w$listTabs} lists the tab names of a \code{tabsetPanel} widget.
-#' It fails for other types of widgets.
-#'
-#' @name Widget
-#' @examples{
-#'
-#' }
-NULL
-
 #' @importFrom R6 R6Class
-
 Widget <- R6Class(
   "Widget",
-
+  private = list(
+    name = NULL,     # name in shiny
+    element = NULL,  # HTML element with name as id
+    type = NULL,     # e.g. selectInput
+    iotype = NULL    # "input" or "output"
+  ),
   public = list(
-    initialize = function(name, element, type,
-      iotype = c("input", "output"))
-      widget_initialize(self, private, name, element, type,
-                        match.arg(iotype)),
+    #' @description Create new `Widget`
+    #' @param name Name of a Shiny widget.
+    #' @param element [webdriver::Element]
+    #' @param type Widget type
+    #' @param iotype Input/output type.
+    initialize = function(name, element, type, iotype = c("input", "output")) {
+      iotype <- match.arg(iotype)
 
+      private$name <- name
+      private$element <- element
+      private$type <- type
+      private$iotype <- iotype
+      invisible(self)
+    },
+
+    #' @description Control id (i.e. `inputId` or `outputId` that control
+    #'   was created with).
     getName = function() private$name,
+    #' @description Underlying [webdriver::Element()] object.
     getElement = function() private$element,
+    #' @description retrieve the underlying HTML for a widget
+    getHtml = function() {
+      private$element$executeScript("return arguments[0].outerHTML;")
+    },
+    #' @description Widget type, e.g. `textInput`, `selectInput`.
     getType = function() private$type,
+    #' @description Is this an input or output control?
     getIoType = function() private$iotype,
+    #' @description Is this an input control?
     isInput = function() private$iotype == "input",
+    #' @description Is this an output control?
     isOutput = function() private$iotype == "output",
 
-    getValue = function()
-      widget_getValue(self, private),
+    #' @description Get current value of control.
+    getValue = function(){
+      "!DEBUG widget_getValue `private$name`"
 
-    setValue = function(value)
-      widget_setValue(self, private, value),
+      if (private$iotype == "input") {
+        res <- private$element$executeScript(
+          "var el = $(arguments[0]);
+           return el.data('shinyInputBinding').getValue(el[0]);"
+        )
+      } else {
+        res <- switch(private$type,
+          htmlOutput = private$element$executeScript("return $(arguments[0]).html();"),
+          verbatimTextOutput = private$element$getText(),
+          textOutput = private$element$getText(),
+          abort(paste0("getValue is not implemented for ", private$type))
+        )
+      }
 
-    sendKeys = function(keys)
-      widget_sendKeys(self, private, keys),
+      # Post-process, if needed
+      res <- switch(private$type,
+        checkboxGroupInput = as.character(unlist(res)),
+        dateInput = as.Date(res),
+        dateRangeInput = as.Date(unlist(res)),
+        sliderInput = as.numeric(unlist(res)),
+        res
+      )
 
-    listTabs = function()
-      widget_listTabs(self, private),
+      res
+    },
 
-    uploadFile = function(filename)
-      widget_uploadFile(self, private, filename)
+    #' @description Set value of control.
+    #' @param value Value to set for the widget.
+    setValue = function(value) {
+      "!DEBUG widget_setValue `private$name`"
+      if (private$iotype == "output") {
+        abort("Cannot set values of output widgets")
+      }
 
-  ),
+      # Preprocess value
+      value <- switch(private$type,
+        dateRangeInput = list(start = value[1], end = value[2]),
+        radioButtons = if (!is.null(value)) as.character(value),
+        value
+      )
 
-  private = list(
-    name = NULL,                        # name in shiny
-    element = NULL,                     # HTML element with name as id
-    type = NULL,                        # e.g. selectInput
-    iotype = NULL                       # "input" or "output"
+      setValueScript <-"
+        var el = $(arguments[0]);
+        var val = arguments[1];
+        el.data('shinyInputBinding').setValue(el[0], val);
+        el.trigger('change');
+      "
+      private$element$executeScript(setValueScript, value)
+
+      invisible(self)
+    },
+
+    #' @description scrolls the element into view, then clicks the in-view
+    #'   centre point of it.
+    #' @return self, invisibly.
+    click = function() {
+      private$element$click()
+      invisible(self)
+    },
+
+    #' @description Send specified key presses to control.
+    #' @param keys Keys to send to the widget or the app. See [webdriver::key]
+    #'   for how to specific special keys.
+    sendKeys = function(keys) {
+      "!DEBUG widget_sendKeys `private$name`"
+      private$element$sendKeys(keys)
+    },
+
+    #' @description Lists the tab names of a [shiny::tabsetPanel()].
+    #'  It fails for other types of widgets.
+    listTabs = function() {
+      if (private$type != "tabsetPanel") {
+        abort("'listTabs' only works for 'tabsetPanel' Widgets")
+      }
+      tabs <- private$element$findElements("li a")
+      vapply(tabs, function(t) t$getData("value"), "")
+    },
+
+    #' @description Upload a file to a [shiny::fileInput()].
+    #'  It fails for other types of widgets.
+    #' @param filename Path to file to upload
+    uploadFile = function(filename) {
+      private$element$uploadFile(filename = filename)
+    }
   )
 )
-
-widget_initialize <- function(self, private, name, element, type, iotype) {
-  private$name <- name
-  private$element <- element
-  private$type <- type
-  private$iotype <- iotype
-  invisible(self)
-}
-
-widget_sendKeys <- function(self, private, keys) {
-  "!DEBUG widget_sendKeys `private$name`"
-  private$element$sendKeys(keys)
-}
-
-widget_listTabs <- function(self, private) {
-  if (private$type != "tabsetPanel") {
-    stop("'listTabs' only works for 'tabsetPanel' Widgets")
-  }
-  tabs <- private$element$findElements("li a")
-  vapply(tabs, function(t) t$getData("value"), "")
-}
-
-widget_uploadFile <- function(self, private, filename) {
-  private$element$uploadFile(
-    filename = filename
-  )
-}