+++ tests/testthat/test-app-set-inputs.R
@@ -1,5 +1,3 @@
-context("app$setInputs")
-
 app <- ShinyDriver$new(test_path("apps/081-widgets-gallery"))
 
 test_that("app$setInputs for all input widgets", {
@@ -69,4 +67,11 @@ test_that("app$setInputs for all input widgets", {
 test_that("app$uploadFile for file inputs", {
   x <- app$uploadFile(file = test_path("apps/081-widgets-gallery/DESCRIPTION"))
   expect_true(grepl("DESCRIPTION", x$output$fileOut))
+})
+
+test_that("numeric input recovers from receiving bad input", {
+  num <- app$findWidget("num")
+  num$setValue("bogus")
+  num$setValue(10)
+  expect_equal(num$getValue(), 10)
 })